% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{add_mito_percentage}
\alias{add_mito_percentage}
\title{Add Percentage of Mitochondrial Genes}
\usage{
add_mito_percentage(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
The updated Seurat object with an additional column for mitochondrial gene percentage.
}
\description{
Adds a new column to the Seurat object that contains the percentage of mitochondrial genes.
This is used for quality control purposes.
}
