% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_aggregate.R
\name{get_data_aggregate}
\alias{get_data_aggregate}
\title{Get data aggregate}
\usage{
get_data_aggregate(
  system_id = as.numeric(Sys.getenv("TIGO_ID")),
  datetime_min = Sys.time() - 7 * 24 * 3600,
  datetime_max = Sys.time(),
  level = "min",
  sensors = TRUE,
  params = c("Pin", "Vin", "Iin", "RSSI"),
  header = "key",
  tz = ""
)
}
\arguments{
\item{system_id}{ID of the system (integer), (default: as.numeric(Sys.getenv("TIGO_ID")))}

\item{datetime_min}{Start datetime of data (default: Sys.time() - 7\emph{24}3600)}

\item{datetime_max}{End datetime of data (default: Sys.time())}

\item{level}{"min" (minute ) or "day" (aggregates), default: "min"}

\item{sensors}{Whether to include sensor data TRUE/FALSE (default: TRUE)}

\item{params}{Type of data to return (default: c("Pin", "Vin", "Iin", "RSSI"))}

\item{header}{either "id" or "key" header (default: "key")}

\item{tz}{time zone of system (default: "")}
}
\value{
list with data
}
\description{
Get data aggregate
}
