% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{get_earliest_principal_node}
\alias{get_earliest_principal_node}
\title{Get Earliest Principal Node}
\usage{
get_earliest_principal_node(cds, seurat_cluster, start_cluster)
}
\arguments{
\item{cds}{A Monocle3 cell data set (CDS) object.}

\item{seurat_cluster}{The name of the cluster column in the Seurat object.}

\item{start_cluster}{The starting cluster for the trajectory.}
}
\value{
The earliest principal node in the trajectory graph.
}
\description{
This function identifies the earliest principal node for ordering cells in a Monocle3 trajectory. It calculates the root principal node based on the start cluster.
}
