% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigo-api.R
\name{get_token}
\alias{get_token}
\title{Get token}
\usage{
get_token()
}
\value{
Oauth token ()
}
\description{
Get token
}
\examples{
\dontrun{
Sys.setenv(TIGO_EMAIL = "myemail")
Sys.setenv(TIGO_PW = "mypw")
get_token()
}
}
