% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{goserver}
\alias{goserver}
\title{GO Enrichment Analysis Server Module}
\usage{
goserver(input, output, session)
}
\arguments{
\item{input}{The input object from the Shiny server function.}

\item{output}{The output object from the Shiny server function.}

\item{session}{The session object from the Shiny server function.}
}
\value{
No direct return value. The function produces reactive outputs in the Shiny app.
}
\description{
This function implements the server logic for the GO Enrichment Analysis module.
}
