% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{goUI}
\alias{goUI}
\title{GO Enrichment Analysis UI Module}
\usage{
goUI(id)
}
\arguments{
\item{id}{A unique identifier for the module's namespace.}
}
\value{
A Shiny UI layout for the GO Enrichment Analysis module.
}
\description{
This function creates the user interface for the GO Enrichment Analysis module.
}
