% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{perform_quality_control}
\alias{perform_quality_control}
\title{Perform Quality Control on Seurat Object}
\usage{
perform_quality_control(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
The filtered Seurat object after applying quality control.
}
\description{
Applies quality control steps to the Seurat object, including adding mitochondrial gene percentages,
and filtering cells based on feature counts and mitochondrial gene percentage.
}
