% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Tigo data}
\usage{
plot_data(data, param = "RSSI", interactive = FALSE)
}
\arguments{
\item{data}{as returned by \code{\link{get_data_aggregate}}}

\item{param}{parameter to plot, one of: "Pin", "Vin", "Iin", "RSSI"
(default: "RSSI")}

\item{interactive}{if FALSE (ggplot2), if TRUE (interactive plot),
(default: FALSE)}
}
\value{
plot selected parameter
}
\description{
Plot Tigo data
}
