% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-string.R
\name{plot_string_sum}
\alias{plot_string_sum}
\title{Plot string sum}
\usage{
plot_string_sum(data, para = "Vin", ypara = "sum", interactive = TRUE)
}
\arguments{
\item{data}{as returned by \code{\link{get_data_aggregate}}}

\item{para}{parameter to plot (default: "Vin")}

\item{ypara}{either "sum" (for parameter selected in argument para
or "n" (number of panels), (default: "param")}

\item{interactive}{TRUE/FALSE}
}
\value{
string sum plot
}
\description{
Plot string sum
}
