% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{processInputData}
\alias{processInputData}
\title{Process Input Data for Trajectory Analysis}
\usage{
processInputData(
  gene_expression = NULL,
  cell_metadata = NULL,
  gene_annotation = NULL,
  seurat_object = NULL
)
}
\arguments{
\item{gene_expression}{A gene expression matrix (default is NULL).}

\item{cell_metadata}{A data frame containing cell metadata (default is NULL).}

\item{gene_annotation}{A data frame containing gene annotations (default is NULL).}

\item{seurat_object}{A Seurat object containing RNA assay data (default is NULL).}
}
\value{
A list containing gene expression, cell metadata, and gene annotation.
}
\description{
This function processes the input data for trajectory analysis. It accepts either a Seurat object or raw data and extracts the gene expression matrix, cell metadata, and gene annotations.
}
