% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{runDEGtrajectory_M}
\alias{runDEGtrajectory_M}
\title{Run Differential Expression Analysis Using Monocle3}
\usage{
runDEGtrajectory_M(
  cds,
  interested_genes,
  seurat_cluster,
  start_cluster,
  q,
  session
)
}
\arguments{
\item{cds}{A Monocle3 cell data set (CDS) object.}

\item{interested_genes}{A vector of gene names of interest.}

\item{seurat_cluster}{The name of the cluster column in the Seurat object.}

\item{start_cluster}{The starting cluster for the trajectory.}

\item{q}{The threshold for the q-value to identify significant genes.}

\item{session}{The Shiny session object.}
}
\value{
A list containing the aggregated expression matrix, heatmap, pseudotime plot, and DEG results.
}
\description{
This function performs differential expression analysis along a trajectory using Monocle3. It identifies differentially expressed genes (DEGs) and generates a heatmap and pseudotime plot.
}
