% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{runDEGtrajectory_S}
\alias{runDEGtrajectory_S}
\title{Run Differential Expression Analysis Using Slingshot}
\usage{
runDEGtrajectory_S(output_slingshot, Tcells, q, session)
}
\arguments{
\item{output_slingshot}{The output from the Slingshot trajectory analysis.}

\item{Tcells}{A Seurat object containing RNA assay data.}

\item{q}{The threshold for the q-value to identify significant genes.}

\item{session}{The Shiny session object.}
}
\value{
A list containing the DEG results, heatmap, and smooth terms.
}
\description{
This function performs differential expression analysis along a trajectory using Slingshot. It fits a GAM model for each gene and generates a heatmap of the expression patterns.
}
