% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_2446312.R
\name{server}
\alias{server}
\title{TIgo Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input, }{output, session Standard Shiny server arguments.}
}
\description{
Defines the server logic for the TIgo Shiny app. It includes server functions for each of the three modules corresponding to the tabs in the UI.
}
