% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigo-api.R
\name{tigo_oauth}
\alias{tigo_oauth}
\title{Tigo oauth endpoints}
\usage{
tigo_oauth(endpoint, token = get_token()$auth)
}
\arguments{
\item{endpoint}{Tigo endpoint}

\item{token}{Tigo token \code{\link{get_token}}}
}
\value{
data returned from endpoint
}
\description{
Tigo oauth endpoints
}
\examples{
\dontrun{
systems <- tigo_oauth(endpoint = tigo_api()$systems)
system_id <- systems$systems[[1]]$system_id
sources <- tigo_oauth(endpoint = sprintf("\%s\%d", tigo_api()$sources, system_id))
objects <- tigo_oauth(endpoint = sprintf("\%s\%d", tigo_api()$objects, system_id))
}
}
