% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{TIserver}
\alias{TIserver}
\title{Trajectory Analysis Module Server}
\usage{
TIserver(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server function parameters.}
}
\value{
None. This function is used to handle reactive expressions and events in the module.
}
\description{
This function defines the server logic for the Trajectory Analysis module in a Shiny app.
}
