% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TI.R
\name{TIUI}
\alias{TIUI}
\title{Trajectory Analysis Module UI}
\usage{
TIUI(id)
}
\arguments{
\item{id}{A string representing the namespace ID for the module.}
}
\value{
A Shiny UI definition for the Trajectory Analysis module.
}
\description{
This function defines the UI for the Trajectory Analysis module in a Shiny app.
}
