% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_iso_datetime}
\alias{to_iso_datetime}
\title{Helper function: convert to ISO datetime}
\usage{
to_iso_datetime(datetime)
}
\arguments{
\item{datetime}{date time string as returned by Sys.time()}
}
\value{
reformats to ISO datetime
}
\description{
Helper function: convert to ISO datetime
}
\examples{
to_iso_datetime(Sys.time())

}
