% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_2446312.R
\docType{data}
\name{ui}
\alias{ui}
\title{TIgo UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
Creates the user interface for the TIgo Shiny app using a `navbarPage` layout. The UI consists of three tabs:
- Seurat object creation
- Trajectory Analysis
- GO Enrichment Analysis
}
\keyword{datasets}
