% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{umapserver}
\alias{umapserver}
\title{Define Server Logic for the UMAP Module}
\usage{
umapserver(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\description{
This function defines the server logic for the UMAP module in a Shiny app.
It handles file processing, UMAP computation, clustering, subsetting, and provides
download handlers for Seurat objects and UMAP plots.
}
