% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{umapui}
\alias{umapui}
\title{Define the UMAP Module UI}
\usage{
umapui(id)
}
\arguments{
\item{id}{A unique identifier for the module. It is used to namespace the input and output IDs.}
}
\value{
A `shiny.tag` object representing the UI for the UMAP module.
}
\description{
This function creates the user interface for the UMAP module of the Shiny app,
including file inputs, text inputs, sliders, buttons, and output elements.
}
